switched to using end of base64 encoded random instead of beginning

+++ jenkins/src/com/zevrant/services/services/VersionService.groovy
@@ -66,7 +66,7 @@ class VersionService extends Service {
     private static int generateRandomBuildForVersionForLength(int length) {
         String buildVersion = ""
         for (int i = 0; i < length; i++) {
-            buildVersion += Integer.valueOf(Math.round(Double.valueOf(Math.random() * 10))).toString()
+            buildVersion += Integer.valueOf((int) Math.round(Double.valueOf(Math.random() * 10))).toString()
         }
         return buildVersion
     }
